+++ app/Http/Controllers/NewsController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use App\Models\News;
 use App\Models\NewsImage;
+use App\Models\RolesPermission;
 use App\Models\Seo;
 use App\Models\Tags;
 use App\Models\Setting;
 use Illuminate\Http\Request;
 use DataTables;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
@@ -22,76 +24,85 @@ class NewsController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->ajax()) {
-            $data = News::latest()->get();
-            return Datatables::of($data)
-                ->addIndexColumn()
-                ->addColumn('category', function ($row) {
-                    $categories = $row->category_id;
-                    $categorys = '';
-                    foreach ($categories as $category) {
-                        $category_name = Category::where('id', $category)->first();
-                        $categorys .= '<span class="badge bg-green">' . $category_name->title . '</span>' . ' ';
-                    }
-                    return $categorys;
-                })
-                ->addColumn('featured', function ($row) {
-                    $featured = $row->featured;
-                    if ($featured == 1) {
-                        $featured = "Featured";
-                    } else {
-                        $featured = "Not Featured";
-                    }
-                    return $featured;
-                })
-                // ->addColumn('status', function ($row) {
-                //     $status = $row->status;
-                //     if ($status == 1) {
-                //         $status = "Approved";
-                //     } else {
-                //         $status = "Not Approved";
-                //     }
-                //     return $status;
-                // })
+        $roles_permission = RolesPermission::where('role_id', Auth::user()->role_id)->get();
+        $rolespermission = [];
+        foreach ($roles_permission as $rolepermission) {
+            array_push($rolespermission, $rolepermission->permission_id);
+        }
+        if (in_array(5, $rolespermission)) {
+            if ($request->ajax()) {
+                $data = News::latest()->get();
+                return Datatables::of($data)
+                    ->addIndexColumn()
+                    ->addColumn('category', function ($row) {
+                        $categories = $row->category_id;
+                        $categorys = '';
+                        foreach ($categories as $category) {
+                            $category_name = Category::where('id', $category)->first();
+                            $categorys .= '<span class="badge bg-green">' . $category_name->title . '</span>' . ' ';
+                        }
+                        return $categorys;
+                    })
+                    ->addColumn('featured', function ($row) {
+                        $featured = $row->featured;
+                        if ($featured == 1) {
+                            $featured = "Featured";
+                        } else {
+                            $featured = "Not Featured";
+                        }
+                        return $featured;
+                    })
+                    // ->addColumn('status', function ($row) {
+                    //     $status = $row->status;
+                    //     if ($status == 1) {
+                    //         $status = "Approved";
+                    //     } else {
+                    //         $status = "Not Approved";
+                    //     }
+                    //     return $status;
+                    // })
 
-                ->addColumn('is_trending', function ($row) {
-                    $is_trending = $row->is_trending;
-                    if ($is_trending == 1) {
-                        $is_trending = "Trending";
-                    } else {
-                        $is_trending = "Not Trending";
-                    }
-                    return $is_trending;
-                })
-                ->addColumn('image', function ($row) {
-                    if ($row->image == 'post.jpg') {
-                        $imagename = Storage::disk('uploads')->url('noimage.jpg');
-                    } else {
-                        $imagename = Storage::disk('uploads')->url($row->image);
-                    }
-                    $image = "<img src='$imagename' style='max-height: 100px;'>";
-                    return $image;
-                })
-                ->addColumn('action', function ($row) {
+                    ->addColumn('is_trending', function ($row) {
+                        $is_trending = $row->is_trending;
+                        if ($is_trending == 1) {
+                            $is_trending = "Trending";
+                        } else {
+                            $is_trending = "Not Trending";
+                        }
+                        return $is_trending;
+                    })
+                    ->addColumn('image', function ($row) {
+                        if ($row->image == 'post.jpg') {
+                            $imagename = Storage::disk('uploads')->url('noimage.jpg');
+                        } else {
+                            $imagename = Storage::disk('uploads')->url($row->image);
+                        }
+                        $image = "<img src='$imagename' style='max-height: 100px;'>";
+                        return $image;
+                    })
+                    ->addColumn('action', function ($row) {
 
-                    $editurl = route('admin.news.edit', $row->id);
-                    $deleteurl = route('admin.news.destroy', $row->id);
+                        $editurl = route('news.edit', $row->id);
+                        $deleteurl = route('news.destroy', $row->id);
 
-                    $csrf_token = csrf_token();
+                        $csrf_token = csrf_token();
 
-                    $btn = "<a href='$editurl' class='edit btn btn-primary btn-sm' style='margin-top: 3px;'>Edit</a>
+                        $btn = "<a href='$editurl' class='edit btn btn-primary btn-sm' style='margin-top: 3px;'>Edit</a>
                                 <form action='$deleteurl' method='POST' style='display:inline-block; margin-top: 3px;'>
                                     <input type='hidden' name='_token' value='$csrf_token'>
                                     <input type='hidden' name='_method' value='DELETE' />
                                         <button type='submit' class='btn btn-danger btn-sm'>Delete</button>
                                 </form>";
-                    return $btn;
-                })
-                ->rawColumns(['status', 'featured', 'image', 'category', 'action'])
-                ->make(true);
+                        return $btn;
+                    })
+                    ->rawColumns(['status', 'featured', 'image', 'category', 'action'])
+                    ->make(true);
+            }
+            $setting = Setting::first();
+            return view('backend.news.index', compact('setting'));
+        } else {
+            return view('backend.permissions.permission');
         }
-        $setting = Setting::first();
-        return view('backend.news.index', compact('setting'));
     }
 
     /**
@@ -101,10 +112,19 @@ class NewsController extends Controller
      */
     public function create()
     {
-        $categories = Category::all();
-        $images = NewsImage::where('news_id', 0)->get();        
-        $setting = Setting::first();
-        return view('backend.news.create', compact('categories', 'images', 'setting'));
+        $roles_permission = RolesPermission::where('role_id', Auth::user()->role_id)->get();
+        $rolespermission = [];
+        foreach ($roles_permission as $rolepermission) {
+            array_push($rolespermission, $rolepermission->permission_id);
+        }
+        if (in_array(5, $rolespermission)) {
+            $categories = Category::all();
+            $images = NewsImage::where('news_id', 0)->get();
+            $setting = Setting::first();
+            return view('backend.news.create', compact('categories', 'images', 'setting'));
+        } else {
+            return view('backend.permissions.permission');
+        }
     }
 
     /**
@@ -116,17 +136,17 @@ class NewsController extends Controller
     public function store(Request $request)
     {
         if ($request->ajax()) {
-            $this->validate($request,[
-                'file'=>'required|max:500'
+            $this->validate($request, [
+                'file' => 'required|max:500'
             ]);
-            $name = $request->file->store('newsdetails_images','uploads');
+            $name = $request->file->store('newsdetails_images', 'uploads');
             $i = new NewsImage;
             $i->images = $name;
             $i->news_id = 1;
             $i->title = '';
             $i->save();
 
-            return response()->json(['url'=>Storage::disk('uploads')->url($name),'id'=>$i->id]);
+            return response()->json(['url' => Storage::disk('uploads')->url($name), 'id' => $i->id]);
         };
 
         $data = $this->validate($request, [
@@ -169,11 +189,11 @@ class NewsController extends Controller
         ]);
 
         $images = NewsImage::where('news_id', 1)->get();
-            foreach($images as $image){
-                $image->title = $data['title'];
-                $image->news_id = $news['id'];
-                $image->save();
-            }
+        foreach ($images as $image) {
+            $image->title = $data['title'];
+            $image->news_id = $news['id'];
+            $image->save();
+        }
 
         foreach ($tags as $tag) {
             $tag_info = Tags::create([
@@ -191,11 +211,11 @@ class NewsController extends Controller
             'seodescription' => $data['seodescription'],
         ]);
         $seo_info->save();
-        
+
         $category = Category::where('id', $news->category_id[0])->first();
         FrontController::sendNews($news, $category);
 
-        return redirect()->route('admin.news.index')->with('success', 'News information saved successfully.');
+        return redirect()->route('news.index')->with('success', 'News information saved successfully.');
     }
 
     /**
@@ -217,16 +237,25 @@ class NewsController extends Controller
      */
     public function edit($id)
     {
-        $news = News::findorfail($id);
-        $categories = Category::all();
-        $tags_info = Tags::where('news_id', $id)->get();
-        $setting = Setting::first();
-        $seo_info = Seo::where('news_id', $news->id)->first();
-        $tags = '';
-        foreach ($tags_info as $tag) {
-            $tags .= $tag->tags.',';
+        $roles_permission = RolesPermission::where('role_id', Auth::user()->role_id)->get();
+        $rolespermission = [];
+        foreach ($roles_permission as $rolepermission) {
+            array_push($rolespermission, $rolepermission->permission_id);
         }
-        return view('backend.news.edit', compact('news', 'tags', 'categories', 'seo_info', 'setting'));
+        if (in_array(5, $rolespermission)) {
+            $news = News::findorfail($id);
+            $categories = Category::all();
+            $tags_info = Tags::where('news_id', $id)->get();
+            $setting = Setting::first();
+            $seo_info = Seo::where('news_id', $news->id)->first();
+            $tags = '';
+            foreach ($tags_info as $tag) {
+                $tags .= $tag->tags . ',';
+            }
+            return view('backend.news.edit', compact('news', 'tags', 'categories', 'seo_info', 'setting'));
+        } else {
+            return view('backend.permissions.permission');
+        }
     }
 
     /**
@@ -302,7 +331,7 @@ class NewsController extends Controller
             'seodescription' => $data['seodescription'],
         ]);
 
-        return redirect()->route('admin.news.index')->with('success', 'News information updated successfully.');
+        return redirect()->route('news.index')->with('success', 'News information updated successfully.');
     }
 
     /**
@@ -313,15 +342,24 @@ class NewsController extends Controller
      */
     public function destroy($id)
     {
-        $news = News::findorFail($id);
-        Storage::disk('uploads')->delete($news->image);
-        $tags = Tags::where('news_id', $news->id)->get();
-        foreach ($tags as $tag) {
-            $tag->delete();
+        $roles_permission = RolesPermission::where('role_id', Auth::user()->role_id)->get();
+        $rolespermission = [];
+        foreach ($roles_permission as $rolepermission) {
+            array_push($rolespermission, $rolepermission->permission_id);
         }
-        $seo_info = Seo::where('news_id', $news->id)->first();
-        $seo_info->delete();
-        $news->delete();
-        return redirect()->back()->with('success', 'News information deleted successfully.');
+        if (in_array(5, $rolespermission)) {
+            $news = News::findorFail($id);
+            Storage::disk('uploads')->delete($news->image);
+            $tags = Tags::where('news_id', $news->id)->get();
+            foreach ($tags as $tag) {
+                $tag->delete();
+            }
+            $seo_info = Seo::where('news_id', $news->id)->first();
+            $seo_info->delete();
+            $news->delete();
+            return redirect()->back()->with('success', 'News information deleted successfully.');
+        } else {
+            return view('backend.permissions.permission');
+        }
     }
 }
